requirements.txt
# requirements.txt
# Emoji Transformer — Streamlit deployment
# Install with: pip install -r requirements.txt

# ── Inference (CPU-only — smaller image, Hugging Face Spaces has no GPU) ──────
tensorflow-cpu>=2.13.0

# ── Web dashboard ─────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Numerics / data ───────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0

# ── Training only (not needed at inference time) ──────────────────────────────
# datasets>=2.14.0        # HuggingFace TweetEval loader
# scikit-learn>=1.3.0     # compute_class_weight
# matplotlib>=3.7.0       # training plots
# seaborn>=0.12.0         # training plots
# tqdm>=4.65.0            # progress bars
